feat: integrate confirm drep registration component
